+++ doc/whatsnew.rst
@@ -8,8 +8,8 @@ What's new?
    :backlinks: none
    :depth: 1
 
-Next release
-============
+v2.0.1 (2021-01-31)
+===================
 
 Bug fixes
 ---------